Initial support for 8mu2040 devices.
- we now detect things that also begin 8mu in the MIDI detection
  routines. At some point, we should work out what to do about legacy
  SAMD 8mu devices.
- configuration arrays vary between devices, so are now extracted to
  more generic code.
